+++ config/importmap.rb
@@ -1,4 +1,0 @@
-# Pin npm packages by running ./bin/importmap
-
-pin "rails_clipboard_helper", to: "rails_clipboard_helper.js"
-