+++ lib/rails/clipboard/helper.rb
@@ -3,16 +3,15 @@
 require_relative "helper/version"
 require_relative "helper/view_helpers"
 
-if defined?(Rails::Engine)
-  require_relative "helper/engine"
-elsif defined?(Rails::Railtie)
-  require_relative "helper/railtie"
-end
-
 module Rails
   module Clipboard
     module Helper
       class Error < StandardError; end
+
+      # Automatically include view helpers in ActionView
+      if defined?(ActionView::Base)
+        ActionView::Base.include ViewHelpers
+      end
     end
   end
 end